Remove placeholder (#250)
* Update the type of `ValidatorIndex` from `uint24` to `uint64`

* Update `SlashableAttestation` and `verify_slashable_attestation`

* Remove placeholders

* Remove `latest_vdf_outputs`, `persistent_committees`, `persistent_committee_reassignments`, and `custody_challenges` from `BeaconState`
* Remove `custody_reseeds`, `custody_challenges`,  and `custody_responses` from `BaseBeaconBlock`

* Remove `custody_commitment`

* Remove `vote_count` (we can just use `len(slashable_attestation.validator_indices)`)

* Fix `verify_bitfield`

* Rename `verify_slashable_attestation` to `validate_slashable_attestation` to imply that it's not as same as the spec `verify_slashable_attestation`

* Add more test cases for `test_validate_slashable_attestation`

* Add tests for new SlashableAttestation properties

* `verifiy_bitfield` ->  `validate_bitfield`

* Add tests for `validate_bitfield`

* Refactor

* Apply suggestions from code review


* PR feedback, thanks Nic!

+++ eth2/beacon/types/proposer_slashings.py
@@ -3,7 +3,7 @@ from rlp.sedes import (
     binary,
 )
 from eth2.beacon.sedes import (
-    uint24,
+    uint64,
 )
 from .proposal_signed_data import ProposalSignedData
 from eth2.beacon.typing import (
@@ -16,7 +16,7 @@ from eth2.beacon.constants import EMPTY_SIGNATURE
 class ProposerSlashing(rlp.Serializable):
     fields = [
         # Proposer index
-        ('proposer_index', uint24),
+        ('proposer_index', uint64),
         # First proposal data
         ('proposal_data_1', ProposalSignedData),
         # First proposal signature
